VM: merge full_collector.hpp into full_collector.cpp
